# swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kasir Laundry Pro",
    "version": "1.0.0",
    "description": "API Routes and schema details of Kasir Laundry Pro Services"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/service": {
      "get": {
        "summary": "Get all services",
        "description": "Retrieve a list of all services",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "Filter data by name"
          },
          {
            "in": "query",
            "name": "duration",
            "schema": {
              "type": "string"
            },
            "description": "Filter data by duration"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceAll"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new service",
        "description": "Create a new service record",
        "tags": [
          "Services"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Layanan berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          }
        }
      }
    },
    "/service/{id}": {
      "put": {
        "summary": "Update a service",
        "description": "Update an existing service record",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          },
          "404": {
            "description": "Layanan tidak ditemukan"
          }
        }
      },
      "delete": {
        "summary": "Delete a service",
        "description": "Delete an existing service record",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Service deleted successfully"
          },
          "404": {
            "description": "Layanan tidak ditemukan"
          }
        }
      },
      "get": {
        "summary": "Get a service by ID",
        "description": "Retrieve a service by its ID",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the service to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": "Layanan tidak ditemukan"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceAll": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the service",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "Dry Cleaning"
          }
        }
      },
      "Duration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Duration ID",
            "example": "1"
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes",
            "example": 30
          },
          "duration_name": {
            "type": "string",
            "description": "Name of the duration",
            "example": "Half Hour"
          },
          "price": {
            "type": "number",
            "description": "Price for the duration",
            "example": 10000
          }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the service",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "Dry Cleaning"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measurement",
            "example": "kg"
          },
          "durations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Duration"
            },
            "example": [
              {
                "id": "1",
                "duration": 30,
                "duration_name": "Half Hour",
                "price": 10000
              },
              {
                "id": "2",
                "duration": 60,
                "duration_name": "One Hour",
                "price": 15000
              }
            ]
          }
        }
      },
      "ServiceRequestBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "Dry Cleaning"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measurement",
            "example": "kg"
          },
          "durations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "duration": {
                  "type": "string",
                  "description": "UUID of the duration",
                  "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "price": {
                  "type": "number",
                  "description": "Price for the duration",
                  "example": 10000
                }
              }
            },
            "example": [
              {
                "duration": "550e8400-e29b-41d4-a716-446655440000",
                "price": 10000
              },
              {
                "duration": "550e8400-e29b-41d4-a716-446655440001",
                "price": 15000
              }
            ]
          }
        },
        "required": [
          "name",
          "unit",
          "durations"
        ]
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Layanan berhasil dibuat"
          },
          "data": {
            "$ref": "#/components/schemas/Service"
          }
        },
        "example": {
          "status": "success",
          "message": "Layanan berhasil dibuat",
          "data": {
            "id": "1",
            "name": "Dry Cleaning",
            "unit": "kg",
            "durations": [
              {
                "id": "550e8400-e29b-41d4-a716-446655440000",
                "duration": 30,
                "duration_name": "Half Hour",
                "price": 10000
              },
              {
                "id": "550e8400-e29b-41d4-a716-446655440001",
                "duration": 60,
                "duration_name": "One Hour",
                "price": 15000
              }
            ]
          }
        }
      },
...
    }
  },
...
}